moved banmgr in its package, removed s.FilterDB
